ranger les tests e2e
